Add blogpost on Git and Avengers  Endgame (#252)
Resolves #234

Actually, I just want get this out of the way. Enjoyed writing about it, but I realized that I don't like listicle formats.
I opted to use a listicle because I don't know to best write it without meandering and rambling too much.

Also, treating this blogpost as practice for #247. I still need to rethink how to write that. Well, that's a problem for future me!